Save and functionality (#3)
* Populates some initial values with the save

* Adds Contributing and Updates Readme

---------

+++ Assets/Scripts/Managers/Utility/SaveData.cs
@@ -1,11 +1,10 @@
 using System;
-using UnityEditor;
 
 [Serializable]
 public class SaveData
 {
     // Currencies
-    public float money;
+    public double money;
 
     // Object Upgrades
     public float objectValue;
